middleware added

+++ src/modules/issues/issue.controller.ts
@@ -2,16 +2,22 @@ import type { Request, Response } from "express";
 import sendResponse from "../../utility/sendResponse";
 import { issueService } from "./issue.service";
 
+// CREATE ISSUE
 const createIssue = async (req: Request, res: Response) => {
-
   try {
-    const result = await issueService.createIssueIntoDB(req.body);
+    const reporter_id = req?.user?.id;
+    const result = await issueService.createIssueIntoDB({
+      ...req.body,
+      reporter_id,
+    });
+
     sendResponse(res, {
       statusCode: 201,
       success: true,
-      message: "Issue Created successfully!",
-      data: result.rows[0],
+      message: "Issue created successfully",
+      data: result,
     });
+
   } catch (error: any) {
     sendResponse(res, {
       statusCode: 500,
@@ -22,14 +28,30 @@ const createIssue = async (req: Request, res: Response) => {
   }
 };
 
+// GET ALL ISSUES
 const getAllIssues = async (req: Request, res: Response) => {
   try {
-    const result = await issueService.getAllIssuesFromDB();
+    const { sort, type, status } = req.query;
+    const result = await issueService.getAllIssuesFromDB({
+      sort: sort as string,
+      type: type as string,
+      status: status as string,
+    });
+
+    if (result.length === 0) {
+      sendResponse(res, {
+        statusCode: 404,
+        success: false,
+        message: "No issues found!",
+      });
+      return;
+    }
+
     sendResponse(res, {
       statusCode: 200,
       success: true,
       message: "Issues retrieved successfully!",
-      data: result.rows,
+      data: result,
     });
   } catch (error: any) {
     sendResponse(res, {
@@ -41,10 +63,12 @@ const getAllIssues = async (req: Request, res: Response) => {
   }
 };
 
+// GET SINGLE ISSUE
 const getSingleIssue = async (req: Request, res: Response) => {
   const { id } = req.params;
   try {
     const result = await issueService.getSingleIssueFromDB(id as string);
+
     if (result.rows.length === 0) {
       sendResponse(res, {
         statusCode: 404,
@@ -52,6 +76,7 @@ const getSingleIssue = async (req: Request, res: Response) => {
         message: "Issue Not found!",
         data: {},
       });
+      return;
     }
 
     sendResponse(res, {
@@ -70,10 +95,48 @@ const getSingleIssue = async (req: Request, res: Response) => {
   }
 };
 
+//UPDATE ISSUE
 const updateIssue = async (req: Request, res: Response) => {
   const { id } = req.params;
 
   try {
+    const loggedInUser = req.user;
+    const issueResult = await issueService.getSingleIssueFromDB(id as string);
+
+    if (issueResult.rows.length === 0) {
+      sendResponse(res, {
+        statusCode: 404,
+        success: false,
+        message: "Issue Not found!",
+      });
+      return;
+    }
+
+    const issue = issueResult.rows[0];
+    const isMaintainer = loggedInUser?.role === "maintainer";
+    const isOwner = issue.reporter_id === Number(loggedInUser?.id);
+    const isOpen = issue.status === "open";
+
+    if (!isMaintainer) {
+      if (!isOwner) {
+        sendResponse(res, {
+          statusCode: 403,
+          success: false,
+          message: "You can update only your own issues!",
+        });
+        return;
+      }
+
+      if (!isOpen) {
+        sendResponse(res, {
+          statusCode: 403,
+          success: false,
+          message: "Only open issues can be updated!",
+        });
+        return;
+      }
+    }
+
     const result = await issueService.updateIssueFromDB(req.body, id as string);
 
     if (result.rows.length === 0) {
@@ -82,6 +145,7 @@ const updateIssue = async (req: Request, res: Response) => {
         success: false,
         message: "Issue Not found!",
       });
+      return;
     }
 
     sendResponse(res, {
@@ -100,9 +164,23 @@ const updateIssue = async (req: Request, res: Response) => {
   }
 };
 
+
+// DELETE ISSUE
 const deleteIssue = async (req: Request, res: Response) => {
   const { id } = req.params;
   try {
+    const loggedInUser = req.user;
+    const isMaintainer = loggedInUser?.role === "maintainer";
+
+    if (!isMaintainer) {
+      sendResponse(res, {
+        statusCode: 403,
+        success: false,
+        message: "Only maintainers can delete issues!",
+      });
+      return;
+    }
+
     const result = await issueService.deleteIssueFromDB(id as string);
 
     if (result.rowCount === 0) {
@@ -111,6 +189,7 @@ const deleteIssue = async (req: Request, res: Response) => {
         success: false,
         message: "Issue Not found!",
       });
+      return;
     }
 
     sendResponse(res, {
@@ -127,7 +206,9 @@ const deleteIssue = async (req: Request, res: Response) => {
     });
   }
 };
-    
+
+
+//EXPORT CONTROLLER
 export const issueController = {
   createIssue,
   getAllIssues,
